changes, and delete functions dir, move all files to filtersfile.
